renamed openquakeplatform_plugins/ipt to openquakeplatform_ipt
