+++ MANIFEST.in
@@ -1,5 +1,6 @@
 include LICENSE
 include README.rst
-recursive-include openquakeplatform_plugins *.py
-recursive-include openquakeplatform_plugins/ipt/static *
-recursive-include openquakeplatform_plugins/ipt/templates *
+recursive-include openquakeplatform_ipt *.py
+recursive-include openquakeplatform_ipt/management *
+recursive-include openquakeplatform_ipt/static *
+recursive-include openquakeplatform_ipt/templates *
