+++ openquakeplatform_ipt/static/ipt/css/risk_input_model.css
