+++ openquakeplatform_ipt/static/ipt/js/examples/101.js
