+++ openquakeplatform_ipt/static/ipt/js/examples/198.js
