+++ openquakeplatform_ipt/static/ipt/js/examples/299.js
