+++ openquakeplatform_ipt/static/ipt/js/examples/301.js
