+++ openquakeplatform_ipt/static/ipt/js/examples/399.js
