+++ openquakeplatform_ipt/static/ipt/js/examples/401.js
