+++ openquakeplatform_ipt/static/ipt/js/examples/499.js
