+++ openquakeplatform_ipt/static/ipt/js/exposure.js
