+++ openquakeplatform_ipt/static/ipt/js/fragility.js
