+++ openquakeplatform_ipt/static/ipt/js/selectAllTextareaText.js
