+++ openquakeplatform_ipt/static/ipt/js/selectTextareaLine.js
