+++ openquakeplatform_ipt/static/ipt/js/site_conds.js
