+++ openquakeplatform_ipt/static/ipt/js/validateAndDisplayNRML.js
