+++ openquakeplatform_ipt/static/ipt/js/vulnerability.js
