+++ openquakeplatform_ipt/static/ipt/js/vulnerability_new.js
