+++ openquakeplatform_ipt/templates/ipt/ipt.html
