+++ openquakeplatform_ipt/templates/ipt/tabs/exposure.html
