+++ openquakeplatform_ipt/templates/ipt/tabs/fragility.html
