+++ openquakeplatform_ipt/templates/ipt/tabs/vulnerability.html
