+++ openquakeplatform_ipt/templates/ipt/tabs/vulnerability_new.html
