+++ openquakeplatform_ipt/templates/lang.js
