+++ openquakeplatform_ipt/views.py
